_strtod_l

+++ esphome/components/cover/cover.cpp
@@ -79,13 +79,13 @@ void CoverCall::perform() {
   }
   if (this->position_.has_value()) {
     if (traits.get_supports_position()) {
-      ESP_LOGD(TAG, "  Position: %.0f%%", *this->position_ * 100.0f);
+      ESP_LOGD(TAG, "  Position: %d%%", (int) (*this->position_ * 100.0f));
     } else {
       ESP_LOGD(TAG, "  Command: %s", LOG_STR_ARG(cover_command_to_str(*this->position_)));
     }
   }
   if (this->tilt_.has_value()) {
-    ESP_LOGD(TAG, "  Tilt: %.0f%%", *this->tilt_ * 100.0f);
+    ESP_LOGD(TAG, "  Tilt: %d%%", (int) (*this->tilt_ * 100.0f));
   }
   if (this->toggle_.has_value()) {
     ESP_LOGD(TAG, "  Command: TOGGLE");
@@ -150,7 +150,7 @@ void Cover::publish_state(bool save) {
   ESP_LOGD(TAG, "'%s' >>", this->name_.c_str());
   auto traits = this->get_traits();
   if (traits.get_supports_position()) {
-    ESP_LOGD(TAG, "  Position: %.0f%%", this->position * 100.0f);
+    ESP_LOGD(TAG, "  Position: %d%%", (int) (this->position * 100.0f));
   } else {
     if (this->position == COVER_OPEN) {
       ESP_LOGD(TAG, "  State: OPEN");
@@ -161,7 +161,7 @@ void Cover::publish_state(bool save) {
     }
   }
   if (traits.get_supports_tilt()) {
-    ESP_LOGD(TAG, "  Tilt: %.0f%%", this->tilt * 100.0f);
+    ESP_LOGD(TAG, "  Tilt: %d%%", (int) (this->tilt * 100.0f));
   }
   ESP_LOGD(TAG, "  Current Operation: %s", LOG_STR_ARG(cover_operation_to_str(this->current_operation)));
